refactor: extract agent UI and WebSocket transport into modules
Moves the agent toast/dialog cluster into modules/agent-ui.js and the
outbound WebSocket plumbing into modules/ws-transport.js.

The inbound handleWsMessage switch and the agents/activeAgentId state stay
in app.js: they are read from across the file, so extracting them would
trade a large file for a large context object.
